[Test] Reorganize ops/utils tests to mirror source layout
`tests/ops/test_index.py` and `tests/ops/utils/test_index.py` shared a
basename, so pytest's prepend import mode registered both as module
`test_index` and aborted full-tree collection with an import file mismatch
(local single-file runs never triggered it).

Merge the two `test_index.py` into `tests/ops/utils/test_index.py`, keeping
the parametrized correctness + torch.compile suite and folding in the
chunk-offset / split-cu-seqlens / edge-case coverage. Split the catch-all
`tests/ops/test_utils.py` into per-module files under `tests/ops/utils/`
(test_cumsum, test_pooling, test_pack) mirroring `fla/ops/utils/`.
